requirements.txt
requests>=2.28.2
pandas>=1.3.5
pytz>=2022.7.1
pystray>=0.19.4
beautifulsoup4>=4.11.2
Pillow>=9.4.0